Extract more Accounts logic into context
