+++ lib/habits_web/views/api/v1/account_view.ex
@@ -1,12 +1,12 @@
 defmodule HabitsWeb.API.V1.AccountView do
   use Habits.Web, :view
 
-  alias Habits.{Accounts.Account}
+  alias Habits.{Accounts}
 
   def render("show.json", %{account: account}) do
     %{
       email: account.email,
-      checkInData: Account.check_in_data(account)
+      checkInData: Accounts.time_series_check_in_data(account)
     }
   end
